refactor: move tests around, rename functions for clarity
