+++ packages/repocop/src/remediations/repository-02-branch_protection.test.ts
@@ -3,9 +3,7 @@ import type {
 	repocop_github_repository_rules,
 	view_repo_ownership,
 } from '@prisma/client';
-import type { UpdateMessageEvent } from 'common/types';
-import { createBranchProtectionWarningMessages } from './repository-02-branch_protection';
-import { createSqsEntry } from './shared-utilities';
+import { createBranchProtectionWarningMessageEvents } from './repository-02-branch_protection';
 
 const nullOwner: view_repo_ownership = {
 	full_name: '',
@@ -71,7 +69,7 @@ describe('Team slugs should be findable for every team associated with a repo', 
 			slug: 'team-one',
 		};
 
-		const actual = createBranchProtectionWarningMessages(
+		const actual = createBranchProtectionWarningMessageEvents(
 			[evaluatedRepo],
 			[repoOwner],
 			[githubTeam],
@@ -101,7 +99,7 @@ describe('Team slugs should be findable for every team associated with a repo', 
 			slug: 'team-one',
 		};
 
-		const actual = createBranchProtectionWarningMessages(
+		const actual = createBranchProtectionWarningMessageEvents(
 			[evaluatedRepo],
 			[],
 			[githubTeam],
@@ -109,24 +107,5 @@ describe('Team slugs should be findable for every team associated with a repo', 
 		);
 
 		expect(actual.length).toEqual(0);
-	});
-});
-
-describe('Batch entries should be created for each message', () => {
-	test('The batch ID of the message should contain no special characters', () => {
-		const event1: UpdateMessageEvent = {
-			fullName: 'guardian/repo-1',
-			teamNameSlugs: ['team-one'],
-		};
-		const event2: UpdateMessageEvent = {
-			fullName: '!@£$%^&*()l',
-			teamNameSlugs: ['team-two'],
-		};
-
-		const actual1 = createSqsEntry(event1);
-		const actual2 = createSqsEntry(event2);
-
-		expect(actual1.Id).toEqual('guardianrepo1');
-		expect(actual2.Id).toEqual('l');
 	});
 });
